Move test to more fitting module
